pyproject.toml
[tool.poetry]
name = "list_django"
version = "0.1.0"
description = ""
authors = ["Plinio Schmidt de Arruda Junior <pliniojunior5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
